requirements.txt
Django>=3.2.3,<3.3
djangorestframework>=3.12.0,<3.13.0
flake8>=3.9.2,<4.0.0
model_bakery>=1.3.1,<1.4.0
psycopg2>=2.8.6,<2.9.0
pytest>=6.2.4,<6.3
pytest-django>=4.3.0,<4.4.0
